Moved contact specific receiving functions to new `receive::contact` crate.
